+++ ghg_budget/artifact.py
@@ -4,21 +4,19 @@ from plotly.graph_objects import Figure
 from climatoology.base.artifact import (
     create_markdown_artifact,
     create_table_artifact,
-    create_chart_artifact,
     _Artifact,
-    Chart2dData,
     create_plotly_chart_artifact,
 )
 from climatoology.base.computation import ComputationResources
 
-from ghg_budget.data import now_year
+from ghg_budget.data import NOW_YEAR, AOI_EMISSION_END_YEAR
 
 
 def build_methodology_description_artifact(text: str, resources: ComputationResources) -> _Artifact:
     return create_markdown_artifact(
         text=text,
         name='Berechnung des CO₂-Budgets',
-        tl_dr='Beschreibung der Methodik zur Berechnung des CO₂-Budgets',
+        tl_dr=' ',
         resources=resources,
         filename='methodology_description',
     )
@@ -28,7 +26,7 @@ def build_methodology_description_simple_artifact(text: str, resources: Computat
     return create_markdown_artifact(
         text=text,
         name='Berechnung des CO₂-Budgets',
-        tl_dr='Beschreibung der Methodik zur Berechnung des CO₂-Budgets',
+        tl_dr=' ',
         resources=resources,
         filename='simple_methodology_description',
     )
@@ -59,7 +57,7 @@ def build_budget_table_artifact(table: pd.DataFrame, resources: ComputationResou
         '[BISKO](https://www.kea-bw.de/fileadmin/user_upload/Energiemanagement/Angebote/Beschreibung_der_BISKO-Methodik.pdf) '
         'ist ein vom Institut für Energie- und Umweltforschung (IFEU) entwickelter Standard, nach dem '
         'viele Städte wie beispielsweise Heidelberg ihre Emissionen schätzen.\n\n'
-        f'**BISKO CO₂-Budget {now_year} (1000 Tonnen):** CO₂-Budgets, die Heidelberg aktuell noch zur '
+        f'**BISKO CO₂-Budget {NOW_YEAR} (1000 Tonnen):** CO₂-Budgets, die Heidelberg aktuell noch zur '
         'Verfügung stehen. Ein negativer Wert bedeutet, dass das verfügbare Budget bereits überschritten '
         'ist.\n\n'
         '**CO₂-Budget aufgebraucht (Jahr):** Wann die CO₂-Budgets aufgebraucht sind, hängt davon ab, '
@@ -91,7 +89,7 @@ def build_budget_table_simple_artifact(table: pd.DataFrame, resources: Computati
         '**Erläuterung der Spalten**\n\n'
         '**Temperaturziel (°C):** Angestrebte Begrenzung auf eine maximale Erwärmung. Das internationale Abkommen von '
         'Paris gibt eine Begrenzung auf deutlich unter 2 °C Temperaturerhöhung vor.\n\n'
-        f'**BISKO CO₂-Budget {now_year} (1000 Tonnen):** CO₂-Budgets, die Heidelberg aktuell noch zur Verfügung stehen. Ein '
+        f'**BISKO CO₂-Budget {NOW_YEAR} (1000 Tonnen):** CO₂-Budgets, die Heidelberg aktuell noch zur Verfügung stehen. Ein '
         'negativer Wert bedeutet, dass das verfügbare Budget bereits überschritten ist. '
         '[BISKO](https://www.kea-bw.de/fileadmin/user_upload/Energiemanagement/Angebote/Beschreibung_der_BISKO-Methodik.pdf) '
         'ist ein vom Institut für Energie- und Umweltforschung (IFEU) entwickelter Standard, nach dem '
@@ -112,15 +110,16 @@ def build_budget_table_simple_artifact(table: pd.DataFrame, resources: Computati
     )
 
 
-def build_budget_comparison_chart_artifact(bar_chart_data: Chart2dData, resources: ComputationResources) -> _Artifact:
-    return create_chart_artifact(
-        data=bar_chart_data,
+def build_budget_comparison_chart_artifact(fig: Figure, resources: ComputationResources) -> _Artifact:
+    return create_plotly_chart_artifact(
+        figure=fig,
         title='Wie viel vom CO₂-Budget ist bereits emittiert?',
         caption='CO₂-Budgets für Heidelberg, um verschiedene Temperaturziele mit einer Wahrscheinlichkeit von 83% zu '
         'erreichen.',
         description='Das Diagramm zeigt die CO₂-Budgets, die Heidelberg seit der Pariser Klimakonferenz 2016 zur '
         'Verfügung standen, um verschiedene Temperaturziele mit einer Wahrscheinlichkeit von 83 % zu erreichen. Die '
-        'dunkelgraue Säule daneben zeigt, wieviel CO₂ Heidelberg seit 2016 ausgestoßen hat, d.h. wieviel von '
+        f'dunkelgraue Säule daneben zeigt, wieviel CO₂ Heidelberg von 2016 bis {AOI_EMISSION_END_YEAR}, '
+        'dem aktuellsten Jahr, für das Daten vorliegen, ausgestoßen hat, d.h. wieviel von '
         'diesen Budgets bereits verbraucht ist. Die hellgraue Säule ganz rechts zeigt, wieviel CO₂ Heidelberg '
         'voraussichtlich noch ausstoßen wird, bevor die Klimaneutralität erreicht wird. Diese Prognose basiert '
         'auf den aktuell beschlossenen Klimaschutzmaßnahmen der Stadt Heidelberg. Das Diagramm zeigt, dass '
@@ -137,8 +136,9 @@ def build_time_chart_artifact(line_chart: Figure, resources: ComputationResource
         figure=line_chart,
         title='Entwicklung der CO₂-Emissionen in Heidelberg',
         caption='Entwicklung der CO₂-Emissionen Heidelbergs und alternative Reduktionspfade seit 2016 (in 1000 Tonnen)',
-        description='Die Emissionswerte von 2016 bis 2021 in der Datenreihe "Prognose" sind Messwerte basierend auf '
-        'dem BISKO-Standard, die Werte ab 2022 sind Prognosen unter der Annahme, dass die zurzeit '
+        description=f'Die Emissionswerte von 2016 bis {AOI_EMISSION_END_YEAR} in der Datenreihe '
+        '"Prognose" sind Messwerte basierend auf dem BISKO-Standard, die Werte ab '
+        f'{AOI_EMISSION_END_YEAR + 1} sind Prognosen unter der Annahme, dass die zurzeit '
         'beschlossenen Maßnahmen Heidelbergs zur Emissionsreduzierung umgesetzt werden. In dieser Prognose '
         'wird Heidelberg erst 2068 CO₂-neutral. Die alternativen '
         'Reduktionspfade stellen fiktive Szenarien dar, wie Heidelberg bis 2040 CO₂-Neutralität erreichen '
@@ -155,9 +155,9 @@ def build_time_chart_artifact(line_chart: Figure, resources: ComputationResource
     )
 
 
-def build_cumulative_chart_artifact(line_chart_data: Chart2dData, resources: ComputationResources) -> _Artifact:
-    return create_chart_artifact(
-        data=line_chart_data,
+def build_cumulative_chart_artifact(fig: Figure, resources: ComputationResources) -> _Artifact:
+    return create_plotly_chart_artifact(
+        figure=fig,
         title='Kumulative CO₂-Emissionen in Heidelberg',
         caption='Aufsummierte CO₂-Emissionen Heidelbergs pro Jahr seit 2016 (in 1000 Tonnen)',
         description='Ein Rückgang der CO₂-Emissionen bedeutet nicht, dass die CO₂-Konzentration in der Atmosphäre '
@@ -165,8 +165,9 @@ def build_cumulative_chart_artifact(line_chart_data: Chart2dData, resources: Com
         'wie mehr CO₂ in die Atmosphäre gelangt als aus ihr entweicht. Man kann sich das wie eine Badewanne '
         'vorstellen, in die Wasser eingelassen wird. Wenn ich den Wasserhahn ein Stück zudrehe, läuft zwar weniger '
         'Wasser in die Wanne, aber solange der Stöpsel zu ist, steigt der Wasserstand trotzdem weiter an, wenn auch '
-        'langsamer. Dies wird in diesem Diagramm gezeigt. Die Emissionswerte von 2016 bis 2021 sind Messwerte '
-        'basierend auf dem BISKO-Standard, die Werte ab 2022 sind Prognosen unter der Annahme, dass die zurzeit '
+        'langsamer. Dies wird in diesem Diagramm gezeigt. Die Emissionswerte von 2016 bis '
+        f'{AOI_EMISSION_END_YEAR} sind Messwerte basierend auf dem BISKO-Standard, die Werte ab '
+        f'{AOI_EMISSION_END_YEAR + 1} sind Prognosen unter der Annahme, dass die zurzeit '
         'beschlossenen Maßnahmen Heidelbergs zur Emissionsreduzierung umgesetzt werden. \n\n'
         'Anmerkung: Die Emissionswerte bilden '
         'nicht die gesamten Emissionen der Stadt Heidelberg ab, sondern nur etwa 64 % der Emissionen. Dies liegt '
@@ -186,10 +187,10 @@ def build_emission_reduction_chart_artifact(fig: go.Figure, resources: Computati
         title='CO₂-Emissionsminderungspfade für Heidelberg',
         caption='Auswahl möglicher CO₂-Reduktionspfade für Heidelberg unter Einhaltung des Temperaturgrenzwerts von '
         '+2°C mit 83% Wahrscheinlichkeit',
-        description='Im Jahr 2025 hat die Stadt Heidelberg noch ein CO₂-Budget von etwa 4358 Kilotonnen zur Verfügung, '
+        description='Im Jahr 2025 hat die Stadt Heidelberg noch ein CO₂-Budget von etwa 4175 Kilotonnen zur Verfügung, '
         'um den Temperaturgrenzwert von +2°C mit einer Wahrscheinlihckeit von 83 % einzuhalten. Dieses '
         'Diagramm zeigt mögliche Emissionsminderungspfade. Die Summe der Werte in jeder Kurve entspricht '
-        'etwa dem CO₂-Budget von 4358 Kilotonnen. Das Diagramm zeigt, dass wir mehr Zeit haben, '
+        'etwa dem CO₂-Budget von 4175 Kilotonnen. Das Diagramm zeigt, dass wir mehr Zeit haben, '
         'CO₂-neutral zu werden, wenn wir die Emissionen jedes Jahr um 17 % reduzieren, als wenn wir die '
         'Emissionen linear verringern. Besonders schnell ist das Budget aufgebraucht, wenn wir die '
         'Emissionen gar nicht reduzieren. Dieses Diagramm zeigt lediglich fiktive Szenarien. Eine Prognose '